test(unit): avoid null delete in file-not-found scenario

+++ tests/php/Unit/Controller/AEnvironmentPageAwareControllerTest.php
@@ -92,7 +92,9 @@ final class AEnvironmentPageAwareControllerTest extends TestCase {
 		$nextcloudFile = $root->getFirstNodeById($file->getNodeId());
 		$trashManager = \OCP\Server::get(ITrashManager::class);
 		$trashManager->pauseTrash();
-		$nextcloudFile->delete();
+		if ($nextcloudFile !== null) {
+			$nextcloudFile->delete();
+		}
 
 		$this->expectException(LibresignException::class);
 		$this->expectExceptionCode(404);
